pyproject.toml
[tool.poetry]
name = "karabinerkeyboard"
version = "0.1.0"
description = ""
authors = ["Patrick Cunniff <pjcfifa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
raylib = "=5.0.0.1"
requests = "^2.32.3"
pyyaml = "^6.0.2"

[tool.poetry.dev-dependencies]
pyinstaller = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "karabinerkeyboard"

[tool.poetry.scripts]
bundle-app = "karabinerkeyboard.builder:bundle_app"